Moved the CLASSIFIER_HEADS to its own heads file, changed the reference to it
