+++ flash/core/classification.py
@@ -11,13 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from functools import partial
 from typing import Any, Callable, Dict, List, Mapping, Optional, Sequence, Union
 
 import torch
 import torch.nn.functional as F
 from pytorch_lightning.utilities import rank_zero_warn
-from torch import nn, Tensor
+from torch import Tensor
 from torchmetrics import Accuracy, Metric
 
 from flash.core.adapter import AdapterTask
@@ -352,25 +351,3 @@ class FiftyOneLabelsOutput(ClassificationOutput):
             filepath = sample[DataKeys.METADATA]["filepath"]
             return {"filepath": filepath, "predictions": fo_predictions}
         return fo_predictions
-
-
-CLASSIFIER_HEADS = FlashRegistry("classifier_heads")
-
-
-def _load_linear_head(num_features: int, num_classes: int) -> nn.Module:
-    """Loads a linear head.
-
-    Args:
-        num_features: Number of input features.
-        num_classes: Number of output classes.
-
-    Returns:
-        nn.Module: Linear head.
-    """
-    return nn.Linear(num_features, num_classes)
-
-
-CLASSIFIER_HEADS(
-    partial(_load_linear_head),
-    name="linear",
-)
