+++ flash/text/classification/adapters.py
@@ -21,8 +21,8 @@ import torch
 from torch import Tensor
 
 from flash.core.adapter import Adapter, AdapterTask
-from flash.core.classification import CLASSIFIER_HEADS
 from flash.core.data.io.input import DataKeys
+from flash.core.heads import CLASSIFIER_HEADS
 from flash.core.model import Task
 from flash.core.registry import FlashRegistry
 from flash.core.utilities.imports import _TRANSFORMERS_AVAILABLE
@@ -111,7 +111,6 @@ class GenericCollate:
 
 class GenericAdapter(Adapter):
 
-    # TODO: Move IMAGE_CLASSIFIIER_HEADS out for general classification tasks
     heads: FlashRegistry = CLASSIFIER_HEADS
 
     def __init__(self, backbone, num_classes: int, max_length: int = 128, head="linear"):
